# openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Challenge 3
...
paths:
...
  /user/{userEmail}:
    get:
      description: Get info of individual user
      parameters:
        - in: path
          name: userEmail
          description: email of user
          required: true
          schema:
            type: string
      responses:
        200:
          description: operation successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        
        400:
          description: User is not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
...
  /post/user/{userEmail}:
    get:
      description: Get list of posts belong to user
      parameters:
        - in: path
          name: userEmail
          schema:
            type: string
          required: true
      responses:
        200:
          description: get successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    bodyPost:
                      type: string
                    userName:
                      type: string
        400:
          description: bad request (not exist user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
...
components:
  schemas:
    Response:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        description:
          type: string
...
    User:
      type: object
      properties:
        name:
          type: string
          example: Cuong
        email:
          type: string
          example: nangcuong@gmail.com
          pattern: '^\d{3}-\d{2}-\d{4}$'
        password:
          type: string
          example: cuong123
...